[tetris] Adding does_piece_fit function and implementing the missing features and fixing the problems to make it work
Added features include >= and <= logical operators and multiple return support

+++ middle_end/IR/Value.hpp
@@ -59,7 +59,7 @@ public:
     if (IsFPConst())
       return std::to_string(std::get<double>(Val));
     else
-      return std::to_string(std::get<uint64_t>(Val));
+      return std::to_string((int64_t)std::get<uint64_t>(Val));
   }
 
 private:
